+++ src/main.tsx
@@ -5,12 +5,10 @@ import './index.css';
 import './cloudAuth.css';
 import {cloudSync} from './services/cloudSync';
 import {installDevDiagnostics} from './services/devDiagnostics';
+import {primeLanguagePreferences} from './services/useLanguagePreferences';
 
 installDevDiagnostics();
-void cloudSync.start();
-
-createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-);
+void primeLanguagePreferences().finally(() => {
+  void cloudSync.start();
+  createRoot(document.getElementById('root')!).render(<StrictMode><App /></StrictMode>);
+});